cleanup analysis
