+++ src/Ozds.Business/Analysis/Analysis.cs
@@ -43,3 +43,42 @@ public record MeterAnalysis(
   MeterModel Meter,
   Analysis Analysis
 );
+
+public record Consumption(
+  DateTimeOffset Timestamp,
+  decimal MinActiveEnergy_kWh,
+  decimal MaxActiveEnergy_kWh,
+  decimal ActiveEnergy_kWh,
+  decimal MinReactiveEnergy_kVARh,
+  decimal MaxReactiveEnergy_kVARh,
+  decimal ReactiveEnergy_kVARh,
+  decimal MinApparentEnergy_kVAh,
+  decimal MaxApparentEnergy_kVAh,
+  decimal ApparentEnergy_kVAh
+)
+{
+  public static readonly Consumption Null = new(
+    DateTimeOffset.MinValue,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0
+  );
+}
+
+public record Expenses(
+  DateTimeOffset Timestamp,
+  decimal Total_EUR
+);
+
+public record Load(
+  DateTimeOffset Timestamp,
+  decimal ActivePower_kW,
+  decimal ReactivePower_kVAR,
+  decimal ApparentPower_kVA
+);
