--- src/Ozds.Business/Analysis/Consumption.cs
@@ -1,28 +1,0 @@
-namespace Ozds.Business.Analysis;
-
-public record Consumption(
-  DateTimeOffset Timestamp,
-  decimal MinActiveEnergy_kWh,
-  decimal MaxActiveEnergy_kWh,
-  decimal ActiveEnergy_kWh,
-  decimal MinReactiveEnergy_kVARh,
-  decimal MaxReactiveEnergy_kVARh,
-  decimal ReactiveEnergy_kVARh,
-  decimal MinApparentEnergy_kVAh,
-  decimal MaxApparentEnergy_kVAh,
-  decimal ApparentEnergy_kVAh
-)
-{
-  public static readonly Consumption Null = new(
-    DateTimeOffset.MinValue,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0,
-    0
-  );
-}
