--- src/Ozds.Business/Analysis/Expenses.cs
@@ -1,6 +1,0 @@
-namespace Ozds.Business.Analysis;
-
-public record Expenses(
-  DateTimeOffset Timestamp,
-  decimal Total_EUR
-);
