--- src/Ozds.Business/Analysis/Load.cs
@@ -1,8 +1,0 @@
-namespace Ozds.Business.Analysis;
-
-public record Load(
-  DateTimeOffset Timestamp,
-  decimal ActivePower_kW,
-  decimal ReactivePower_kVAR,
-  decimal ApparentPower_kVA
-);
